+++ frontend/nextjs.tsconfig.json
@@ -1,7 +1,9 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "noEmit": false
+    "noEmit": false,
+    "module": "commonjs"
   },
-  "include": ["next.config.ts"]
+  "include": ["next.config.ts"],
+  "exclude": ["node_modules", "cypress"]
 }
